# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "PedalaMi",
    "description": "PedalaMi: cycling and having fun in your city"
  },
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/teams/create": {
      "post": {
        "summary": "Create new team",
        "tags": [
          "Teams"
        ],
        "description": "Create new team",
        "parameters": [
          {
            "in": "body",
            "name": "Request",
            "description": "Data of the team that we want to create",
            "schema": {
              "properties": {
                "adminId": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Team saved correctly!",
            "schema": {
              "properties": {
                "teamId": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Missing parameters"
          },
          "500": {
            "description": "Error saving the team"
          }
        }
      }
    },
    "/teams/search": {
      "get": {
        "tags": [
          "Teams"
        ],
        "summary": "Get all the teams whose name contains the one provided",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "type": "string",
            "description": "(Part of) the name of the team"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ArrayTeamWithoutJoinedMembers"
            }
          },
          "400": {
            "description": "Missing parameters"
          },
          "500": {
            "description": "Error while searching the teams"
          }
        }
      }
    },
    "/teams/join": {
      "post": {
        "summary": "Join team",
        "tags": [
          "Teams"
        ],
        "description": "Join team",
        "parameters": [
          {
            "in": "body",
            "name": "Request",
            "description": "Team id and id of the user who wants to join",
            "schema": {
              "properties": {
                "teamId": {
                  "type": "string"
                },
                "userId": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Team saved correctly!"
          },
          "400": {
            "description": "Missing parameters"
          },
          "500": {
            "description": "Error saving the team"
          }
        }
      }
    },
    "/teams/leave": {
      "post": {
        "summary": "Leave team",
        "tags": [
          "Teams"
        ],
        "description": "Leave team",
        "parameters": [
          {
            "in": "body",
            "name": "Request",
            "description": "Team id and id of the user who wants to leave",
            "schema": {
              "properties": {
                "teamId": {
                  "type": "string"
                },
                "userId": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Team leaved correctly"
          },
          "400": {
            "description": "Missing parameters"
          },
          "500": {
            "description": "Error saving the team"
          }
        }
      }
    },
    "/teams/getTeam": {
      "get": {
        "tags": [
          "Teams"
        ],
        "summary": "Get all the info of a team given its identifier",
        "parameters": [
          {
            "in": "query",
            "name": "teamId",
            "type": "string",
            "description": "Id of the team"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TeamWJoinedMembers"
            }
          },
          "400": {
            "description": "Missing parameters"
          },
          "500": {
            "description": "Error while searching the teams"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Badge": {
      "properties": {
        "criteria": {
          "type": "string"
        },
        "criteriaValue": {
          "type": "number"
        },
        "image": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "Statistics": {
      "properties": {
        "numberOfRides": {
          "type": "number"
        },
        "totalDuration": {
          "type": "number"
        },
        "totalKm": {
          "type": "number"
        },
        "totalElevationGain": {
          "type": "number"
        },
        "averageSpeed": {
          "type": "number"
        },
        "averageDuration": {
          "type": "number"
        },
        "averageKm": {
          "type": "number"
        },
        "averageElevationGain": {
          "type": "number"
        }
      }
    },
...
    "TeamWJoinedMembers": {
      "properties": {
        "_id": {
          "type": "string"
        },
        "adminId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "members": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BasicUser"
          }
        }
      }
    },
    "TeamWithoutJoinedMembers": {
      "properties": {
        "_id": {
          "type": "string"
        },
        "adminId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "members": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "active_events": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "event_requests": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "BasicUser": {
      "properties": {
        "userId": {
          "type": "string"
        },
        "points": {
          "type": "integer",
          "default": 0
        },
        "badges": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Badge"
          }
        },
        "statistics": {
          "$ref": "#/definitions/Statistics"
        }
      }
    },
...
    "ArrayTeamWithoutJoinedMembers": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/TeamWithoutJoinedMembers"
      }
    }
  }
}